requirements.txt
streamlit>=1.28.0
tensorflow>=2.18.0
pillow>=10.0.0
numpy>=1.26.4
joblib>=1.3.0
scikit-learn>=1.3.0

